requirements.txt
Django==2.2.2
djangorestframework==3.9.4
djangorestframework-bulk==0.2.1
pytz==2019.1
sqlparse==0.3.0
